Prefix branch tag
Branch tag was previously pushed as is. This meant if someone had named a branch as vX.Y.Z, it would be tagged and pushed as such. And it would then be picked up by flux and deployed. The branch tag is mostly used for testing, so renamed with a "build" prefix.

Change-type: minor
Signed-off-by: Stathis Moraitidis <stathis@balena.io>